+++ src/main/java/roomescape/reservation/config/TossClientConfiguration.java
@@ -2,7 +2,6 @@ package roomescape.reservation.config;
 
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.stereotype.Component;
-import roomescape.reservation.infrastructure.TossClientProperties;
 
 @Component
 @EnableConfigurationProperties(TossClientProperties.class)
